--- jweb-console-lang-groovy/src/main/resources/META-INF/services/io.github.pintowar.console.repl.Repl
@@ -1,1 +1,0 @@
-io.github.pintowar.console.repl.impl.GroovyRepl
