+++ jweb-console-lang-groovy/src/main/resources/META-INF/services/io.github.pintowar.jweb.console.repl.Repl
@@ -1,0 +1,1 @@
+io.github.pintowar.jweb.console.repl.impl.GroovyRepl
