--- jweb-console-lang-jruby/src/main/java/io/github/pintowar/console/repl/impl/JRubyRepl.java
@@ -1,22 +1,0 @@
-package io.github.pintowar.console.repl.impl;
-
-import io.github.pintowar.console.repl.ReplJsr223;
-
-import java.util.Map;
-
-public class JRubyRepl extends ReplJsr223 {
-
-    public JRubyRepl() {
-        super();
-    }
-
-    public JRubyRepl(Map<String, Object> bindings) {
-        super(bindings);
-    }
-
-    @Override
-    public String getEngineName() {
-        return "jruby";
-    }
-
-}
