--- jweb-console-spring-boot-starter/src/main/java/io/github/pintowar/console/boot/EvalConsoleHandler.java
@@ -1,54 +1,0 @@
-package io.github.pintowar.console.boot;
-
-import io.github.pintowar.console.repl.Repl;
-import io.github.pintowar.console.repl.ScriptResult;
-import org.springframework.context.ApplicationContext;
-import org.springframework.http.MediaType;
-import org.springframework.web.servlet.function.ServerRequest;
-import org.springframework.web.servlet.function.ServerResponse;
-
-import java.util.Map;
-import java.util.Set;
-
-import static java.util.Collections.singletonMap;
-import static org.springframework.web.servlet.function.ServerResponse.ok;
-
-/**
- * Controller for evaluating scripts from web console.
- */
-public class EvalConsoleHandler {
-
-    private Map<String, Repl> repls;
-    private ApplicationContext applicationContext;
-
-    public EvalConsoleHandler(ApplicationContext applicationContext) {
-        this.applicationContext = applicationContext;
-        this.repls = Repl.getNamedRepls();
-    }
-
-    public ServerResponse index() {
-        return ok().render("/console/index.html");
-    }
-
-
-    public ServerResponse engines() {
-        Set<String> engines = repls.keySet();
-        return json().body(engines);
-    }
-
-    public ServerResponse execute(ServerRequest req) {
-        Repl repl = repls.get(req.pathVariable("engine"));
-        try {
-            String script = req.param("script").orElse(null);
-            ScriptResult result = repl.execute(script, singletonMap("applicationContext", applicationContext));
-            return json().body(result);
-        } catch (RuntimeException e) {
-            return json().body(ScriptResult.create(e));
-        }
-    }
-
-    private ServerResponse.BodyBuilder json() {
-        return ok().contentType(MediaType.APPLICATION_JSON);
-    }
-
-}
