--- samples/boot-console/src/main/java/io/github/pintowar/console/sample/BootJWebConsoleApp.java
@@ -1,12 +1,0 @@
-package io.github.pintowar.console.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BootJWebConsoleApp {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BootJWebConsoleApp.class, args);
-    }
-}
