+++ samples/boot-console/src/main/java/io/github/pintowar/jweb/console/sample/BootJWebConsoleApp.java
@@ -1,0 +1,12 @@
+package io.github.pintowar.jweb.console.sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BootJWebConsoleApp {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BootJWebConsoleApp.class, args);
+    }
+}
