--- samples/mn-console/src/main/java/io/github/pintowar/console/sample/MnJWebConsoleApp.java
@@ -1,11 +1,0 @@
-package io.github.pintowar.console.sample;
-
-
-import io.micronaut.runtime.Micronaut;
-
-public class MnJWebConsoleApp {
-
-    public static void main(String[] args) {
-        Micronaut.run(MnJWebConsoleApp.class, args);
-    }
-}
