+++ samples/mn-console/src/main/java/io/github/pintowar/jweb/console/sample/MnJWebConsoleApp.java
@@ -1,0 +1,11 @@
+package io.github.pintowar.jweb.console.sample;
+
+
+import io.micronaut.runtime.Micronaut;
+
+public class MnJWebConsoleApp {
+
+    public static void main(String[] args) {
+        Micronaut.run(MnJWebConsoleApp.class, args);
+    }
+}
